% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Surv_Diff.R
\name{Cluster_Surv_Diff}
\alias{Cluster_Surv_Diff}
\title{Cluster_Surv_Diff}
\usage{
Cluster_Surv_Diff(MCKV_group)
}
\arguments{
\item{MCKV_group}{the manually grouping results based on MCKNNV_tSNE}
}
\description{
Comparing the survival difference between the group identified by MC-KNNV
}
\details{
Input takes the output of MCKNNV_tSNE
}
