% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_intersection.R
\name{DEG_intersection}
\alias{DEG_intersection}
\title{DEG_intersection}
\usage{
DEG_intersection(RNA_seq_dataframe, clin_info, MCKV_group)
}
\arguments{
\item{RNA_seq_dataframe}{Data frame of the RNA sequencing profiles from public database}

\item{clin_info}{Data frame of the clinical information containing full survival status and survival time.}

\item{MCKV_group}{the manually grouping results based on MCKNNV_tSNE}
}
\description{
DEG analysis and genes intersection for subsequent process
}
\details{
Input takes several data frames which are the exports of MCKNNV and MCKNNV_tSNE
}
