% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCKNNV.R
\name{MCKNNV}
\alias{MCKNNV}
\title{MCKNNV}
\usage{
MCKNNV(expr_dataframe, clin_info, p, n, y)
}
\arguments{
\item{expr_dataframe}{Data frame of the biological input data (e.g. results from ssGSEA)}

\item{clin_info}{Data frame of the clinical information containing full survival status and survival time.}

\item{p}{the proportion of training set to total samples}

\item{n}{the replicated times of kNN algorithm in the MC simulation}

\item{y}{the cutoff year of manual labeling}
}
\description{
Creating MC-KNNV algorithm based on survival status
}
\details{
Input takes two data.frame expr_dataframe and clin_info with three variables p, n, and y
}
