% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCKNNV_tSNE.R
\name{MCKNNV_tSNE}
\alias{MCKNNV_tSNE}
\title{MCKNNV_tSNE}
\usage{
MCKNNV_tSNE(expr_dataframe, clin_info, p, n, y1, y2, y3)
}
\arguments{
\item{expr_dataframe}{Data frame of the biological input data (e.g. results from ssGSEA)}

\item{clin_info}{Data frame of the clinical information containing full survival status and survival time.}

\item{p}{the proportion of training set to total samples}

\item{n}{the replicated times of kNN algorithm in the MC simulation}

\item{y1}{the first cutoff year of manual labeling}

\item{y2}{the second cutoff year of manual labeling}

\item{y3}{the third cutoff year of manual labeling}
}
\description{
Creating tSNE distribution based on the MC-KNNV results
}
\details{
Input takes several data frames which are the exports of MCKNNV with different y
}
