% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Diff.R
\name{Mod_Diff}
\alias{Mod_Diff}
\title{Mod_Diff}
\usage{
Mod_Diff(candidate_gene)
}
\arguments{
\item{candidate_gene}{Data frame of the clinical information and RNA sequencing profiles of the candidate genes selected by WGCNA (top 5 prognostic DEGs in each module)}
}
\description{
Modeling gene selection and model construction based on the DEGs from different groups
}
\details{
Input takes data frame of candidate genes created by the exports of DEG_selection and WGCNA
}
