% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomogram.R
\name{Nomogram}
\alias{Nomogram}
\title{Nomogram}
\usage{
Nomogram(candidate_gene, clin_patho_parameter)
}
\arguments{
\item{candidate_gene}{Data frame of the clinical information and RNA sequencing profiles of the candidate genes selected by WGCNA (top 5 prognostic DEGs in each module)}

\item{clin_patho_parameter}{Data frame of the clinicopathological parameters which are included in nomogram}
}
\description{
Establishing nomogram based on the gene model and clinicopathological parameters
}
\details{
Input takes data frame of candidate genes created by the exports of DEG_selection and WGCNA, with the clinicopathological parameters downloaded from public database
}
